Build and test on OS X in Travis.
Travis supports simultaneous builds and test on Linux and OS X. This change
adds support for that in ycmd. Unfortunately, the OS X infrastructure does not
support the "python" build type, so we are forced to implement multiple python
version installs ourselves.

Multiple Python Versions
========

We use an environment variable YCMD_PYTHON_VERSION in the travis config to set
the major.minor python version to test with.

On linux: We use the ppa fkrull/deadsnakes
(https://launchpad.net/~fkrull/+archive/ubuntu/deadsnakes) to install specific
major.minor python version.
On OSX: Apple ships both python 2.6 and 2.7 as part of the base Yosemite OS
release, as such we don't need to install anything. All we need to do is tell
the OS which "default" python version to use:
http://stackoverflow.com/questions/6998545/how-can-i-make-python-2-6-my-default-in-mac-os-x-lion

Annoyingly, however, Apple doesn't ship libpython2.6.dylib in the same
directory pattern as libpython2.7.dylib (it is only in /usr/lib rather than
also in /python_prefix/lib. A small change is required to build.py to support
this.

Mono/Node/Go
=====

On OS X we just get these from homebrew.

check_output
====

Backport of subprocess.check_output, which is not available on python 2.6.

New travis infrastructure
====

Some changes courtesy of @mispencer, including the csharp config and the
"addons" config which replaces the sudo apt-get etc.

ordereddict is needed on python 2.6 and is not installed by default on linux.
virtualenv is not available on python 2.6 OS X, so we have to install manually.