Pass argv[0] to libclang when we detect a modern libclang.
This was added after 3.7 was released (r253466) so I added some hacks to
not pick it up for older libclangs. This won't work on platforms without
weak symbols (windows) but I guess they can wait until 3.8 is released.

The reason why this is needed comes from standard libraries that are
relative to the compiler and a compilation database is used. There are
some users of this configuration, but the prime example is Xcode, where
the compilation database looks roughly like this:

"command": "/Applications/Xcode.app/[...]/usr/bin/clang

and the standard library is found relatively:
  /Applications/Xcode.app/[...]/usr/bin/../include/c++/v1

Xcode's libclang uses a hack that finds the standard library relatively
to libclang.dylib, but that breaks as soon as another libclang is used.
Passing argv[0] from the compilation database should always work.