RustCompleter is conditionally created
The rust completer hook now checks that the racerd binary has been
built. If not, the RustCompleter is not created. This should allow for
example vim's omnifunc to still function for other rust plugins.

The RustCompleter _logger member was made static in the module since it
is now needed by a static function.