Write racerd logs to file
The utility of having persisted logs is obvious, but it is worth
elaborating on the change for the port specification. With Popen, it is
difficult to achieve the equivalent of a unix `tee` where the input is
available for reading in the calling process AND the stdout/stderr are
written to files. For this reason, the port spec changed to use the
GetUnusedLocalhostPort util.