RustCompleter uses RLock instead of Lock
Several methods had `NoLock` variants which required the caller to hold
a lock as part of the calling contract. There was nothing to prevent
someone from forgetting to acquire such a lock. By replacing Lock with
RLock, it is possible for a single thread to acquire the lock multiple
times, and this removed the need for NoLock variants.