Fix semantic trigger matching
Currently, semantic trigger must match at the column start of the
identifier word under the cursor. This is problematic when the trigger
ends with identifier characters because these characters will always
be after the column start. This is fixed by searching between the
column start and the current cursor column.