Removing ycm_client_support
It was around to supply the FilterAndSortCandidates function to Vim.
This has now been replaced with a /filter_and_sort_candidates handler in
ycmd that behaves the same as the function.

The point of all this is that we don't want to have any ycmd compiled
libs loaded into Vim (through YCM). Loading the lib would mean that the
Python version that YCM uses in Vim and the Python version that ycmd
runs on would have to be the same. So py2 or py3 for both because we
only build ycmd for one Python interpreter.

Since we now don't load the ycm_client_support (or any other) library
into Vim, YCM and ycmd can use completely different Python interpreter
versions. One can be py2, the other can be py3 (or any other
combination). This is the holy grail of Python compatibility for us.

In theory, this new architecture will cause extra latency for YCM
completions that use the OmniCompleter. In practice, it's imperceptible.
The extra latency should be *at most* 5-10ms and the whole thing is
still well below 100ms. I have the YCM part of this change done and
having played with Ruby omnicompletion before and after this change, I
couldn't notice any latency difference.