Fix infinite loop in TypeScript completer
When the TSServer is not running, the _ReaderLoop method will enter
in a infinite loop with no waiting, leading to high cpu usage and
considerable disk space usage from the stderr logfile. We fix this
by using an event that is true if the server is running, false
otherwise. We wait for this event in the _ReaderLoop method.
In addition, we catch the IOError exception from _SendCommand and
_SendRequest raised when no server is running and instead raise a
RuntimeError with an appropriate message.