Update bottle to 0.12.11
Fixes: #670

Currently ycmd is only working with bottle that comes with ycmd.
It is really an old version of bottle and ycmd is not compatible with
recent version.

A Debian package must use system-dependencies instead of third_party modules.

This patch is updating bottle and making ycmd to compatible with latest
bottle.