Add test when no completions are found
Add a test for the Rust completer when no completions are found but the
rust_src_path option is given. In that case, we return an empty list.
Do not use handlers.UpdateUserOptions in UserOption. This replaces the
current ServerState instance by a new one leading to unexpected results
in tests.