Add a long-polling endpoint for asyncronous message delivery
A new API endpoint '/receive_messages' has been added which allows the
server to return messages for display to the user at arbitrary times.

Previously it was only possible for the ycmd server to communicate and
action or update on receipt of a synchronous request from the client.

However, the language server protocol inherently supports asynchronous
message delivery, and it is necessary to provide some of these message
to clients.

In particular, the jdt.ls server takes a very long time to start up and
provides some "status" information. We use a "display message" response
to inform the client to show the user a progress update.

Fully asyncronous delivery of diagnostics.

Implement a proper ready handler (precursor to adding tests)