Support for basic completions
Apply textEdits in completion items

Language server protocol requires that the client can apply textEdits,
rather than just simple text. This is not an optional feature, but ycmd
clients do not have this ability.

The protocol, however, restricts that the edit must include the original
requested completion position, so we can perform some simple text
manipulation to apply the edit to the current line and determine the
completion start column based on that.

In particualr, the jdt.ls server returns textEdits that replace the
entered text for import completions, which is one of the most useful
completions.

TODO: this doesn't currently work where there is a mixture of start
positions.

We also include experimental support for "additionalTextEdits" which
allow automatic insertion of, e.g.,  import statements when selecting
completion items. These are sent on the completion response as an
additional completer data item called 'fixits'. The client applies the
same logic as a standard FixIt once the selected completion item is
inserted.