Allow parallel build of boost
Boost's b2 utility allows for building in parallel, using more jobs in
the same way GNU/make (probably POSIX make) allows - by specifying -j
and the number of cores.

Python's multiprocessing module contains a function that returns the
number of cores on a system. Useful for determining the number of jobs.