Swift Compilation Database Implementation
Implement the clang compilation database standard for swift files.

The format is specified for Clang:
https://clang.llvm.org/docs/JSONCompilationDatabase.html

I think this is a good file format to use for swift.

The implementation includes a "CompilationDatabase" to read in the files
and a high level API to get flags, "Flags". It uses logic based on the
`cpp` version of `Flags`.

I originally hoped to reusing the `cpp` version of this but there 2
show stoppers:

- Flags.py logic is highly specialized to clang based languages. I don't
like the idea of a complicated implementation that works for both swift
and clang.

- Swift users should not have to download the massive payload that
clang is.

The class simply takes a compilation invocation, and converts it to
completion flags.

Additionally, I am caching database loads based on the file timestamp.
Flags are also cached, and I think the invalidation will need more work
in the future. Based on my own testing, the logic seems sufficient for a
first pass.

There is tangentially related work around writing programs to generate
these files for standard build systems:

- SwiftPM / SwiftC. It'd be nice to geneate these comp dbs from
 the output of these programs. Perhaps a proposal can be made to get
 SwiftPM generating these.
- Bazel: it doesn't support comp dbs for any language by default, but
 it is easy to add.
- Xcode: Perhaps we can modify Xcpretty to and similar programs to get
 this working with Xcode. The standard method for comp dbs in
 `xcodebuild` is to parse the output and dump a DB based on that. A few
 OSS solutions exist for this but are limited to `clang` for now. I
 opened an issue here: https://github.com/supermarin/xcpretty/issues/281

I included some assertions that I wrote while developing this, and I'll
need to get it into the YCMD test suite before merging.