Add Unicode support to the filter and sort algorithm
Replace the subsequence tree algorithm by the naive one that consists of going
through the letters of the candidate to find if the query is a subsequence of
that candidate.
Define the query and candidates as a sequence of characters where each
character represents one or more UTF-8 code points and stores the following
properties:
 - the normalized string representing the character;
 - its normalized case-folded version;
 - its normalized case-swapped version;
 - if the character is a letter;
 - if the character is a punctuation;
 - if the character is uppercase.
Write a table of Unicode code points that contains information to split a word
into characters and normalize these characters. Do a binary search on the table
to retrieve a code point.
Store the characters and code points in two distinct repositories similarly to
the candidate repository. This speeds up the initialization of candidates and
reduces memory usage.