Add Unicode support to the filter and sort algorithm
Replace the subsequence tree algorithm by the naive one that consists of going
through the letters of the candidate to find if the query is a subsequence of
that candidate.
Define the query and candidates as a sequence of characters where each
character stores the following properties:
 - the original sequence of bytes representing the character;
 - a sequence of bytes for the uppercase version of that character;
 - a sequence of bytes for the case-swapped version of that character;
 - if the character is a letter;
 - if the character is a punctuation;
 - if the character is uppercase.
Write a table of Unicode characters that are either a letter or a punctuation
as defined by the Unicode consortium. Do a binary search on the table to find
if a string corresponds to one of these characters.
Create a repository to store the characters similarly to the candidate
repository. This speeds up the initialization of candidates and reduces memory
usage.