Update to jdt.ls v0.10.0 and complete coverage tests
This required a number of things, not least implementing a proper model
of the server file state and using that to reduce the number of
didClose/didChange events to send. This fixes the interpretation of the
file_data: only modified buffers are included, so we must read the
contents of non-modified (but open) buffers from disk.