Calculate the start codepoint when we don't resolve itms
Resolving lots of completion items is very slow, so we only do that if
there are 100 or feweer completions.

Most unfortunately Visual Studio Code does some heuristics on completion
items, so that any overlap with existing code is replaced, rather than
appended. This happens to be how jdt.ls returns import completions.

It's common for there to be more than 100 suggestions when typing,
say 'import com.'. This leads to bad behaviour where the insertion ends
up being 'com.com.youcompletme...' which is clearly wrong.

To combad this, when we don't have a textEdit to apply, we detect any
overlap between the insertion_text and the existing text (up to the
completion location) and offset the start_codepoint by that much.

The overlap detection algorithm is efficient (see the link) and simpler
than a dynamic programming one.

Incidentally, we add the 'prefix' key to request_wrap to avoid
recalculating it for each element.