Don't grow the notification queue forever
If the client does not support message polling, or if for some reason
the message poll loop fails, we queue notifications forever without
consuming them. Currently there is no limit to the length of this queue,
which is the definition of a managed memory leak.

Instead we make the queue operate like a ring buffer, overwriting the
oldest message if it wraps. We keep up to a maximum 250 messages. This
number is somewhat arbitrary; we need it to be big enough to hold the
messages delivered reasonably by a server within one message poll
return/retry loop. If we see problems with this number, we can tweak it
easily.