Send the full query to the language server when forced
jdt.ls does not return anything useful for top-level completions,
presumably because with no query to work from it has no way to filter
the almost infinite set of possible suggestions.

However it does provide useful completion at the top level (including
automatic imports) when it is given the query string to work with. The
main drawback of this is that the filtering that gets applied is done by
whatever algorithm jdt.ls applies, not what ycmd applies. This can lead
to confusion and breaking the user's mental model of how ycmd completion
works.

Automatic imports are particularly useful for top-level completions,
which are (in general) forced by the user.

So as a compromise, we use the column_codepoint (i.e. including the
'query') when the user forces semantic completion and use the
start_codepoint (i.e. first char after trigger/beginning of identifier),
when 'triggered'.

This allows for powerful features and a simple mental model: if you
force completion, you might get more relevant results, but the filtering
is necessarily different.