Fix GetDoc command on symbols declared in system headers
The GetDoc command fails to return documentation on symbols declared in system
headers and, in particular, headers included with -isystem. This is due to
libclang ignoring comments from these headers. If a symbol is declared in such
header, get the documentation directly from the corresponding translation unit.
Comments in the main file of a translation unit are not ignored.