Add the regex module to sys.path in ycmd exclusively
Since the regex module has a compiled component, adding it to sys.path in YCM
may cause a segmentation fault if the version of Python that YCM is running on
is different than the one used to compile the module (e.g. YCM running on
Python 3.6 while the regex module is compiled for Python 3.5). Note that the
crash only seems to happen on macOS.