Fix system headers search on macOS
We can't rely on upstream libclang to find the system headers on macOS
because it ignores the -resource-dir flag and use the current working
directory to find the builtin includes instead, which fails since they
are in a different place (in the clang_includes folder). Furthermore,
it's unable to find the framework headers on macOS 10.14 without setting
the sysroot. So, we replicate the directory structure used by upstream
Clang to allow the libclang library to find the builtin includes and we
try to reproduce the logic used by Apple Clang to find the system
headers by looking at the output of the command

  clang++ -x c++ -E -v -

which prints the list of system header directories and by reading the
source code of upstream Clang.