Only resolve filtered candidates
To do this we first calculate the insertion_text based on un-resolved
completion item, then we resolve those that pass through our filter
(throttled by max_num_candidates option).

This isn't perfect as we still perform a lot of calculation on the
pre-filtered candidates, throw most of the data away, then recompute it
all on the filtered set. In the case where there is no resolve required,
and the number of candidates is less than 50, we just end up doing all
this twice.

However, we _need_ to do some of it so that the _filtering_ is correct.
This is a little irksome, so FIXME ony bother with DetailCandidates if
the server requires resolving.