Don't repeat resolve for cached items
We only cache the results of 'ComputeCandidatesInner', which are always
_unresolved_. When returning items, whether from the cache or not, we
always call 'DetailCandidates'. This means that we need to identify
candidates (or at least completion items) which have already been
resolved.

To do this, we store the completion item in the 'extra_data' of the
candidate and set a flag '_resolved' on the item within the cached
completion after resolving it. NOTE: To do this, we have replace the
contents of the item dict within the cache, rather than creating a new
item object.

The effect of this is that we only resolve completions once, even when
the cache returns our "unresolved" candidates to us.