Determine simple subcommands automatically
LSP supports a 'capabilities' machanism where servers and clients can
identify support for certain requests. This implements a general
mechanism for discovery of supported subcommands, and standardises the
commands supported by LSP-based completers (other than Java).

Clangd completer is updated to use the general mechanism, as its
capabilities response is correct.  Java completer is a little more
complicated, so is not migrated in this change.

This will make adding basic completers much easier in future.