Don't read settings on every file parse
We only actually update the server settings when starting the server, so
only read them from the config when starting the server.

This incidentally fixes a bug in the clangd completer where settings
were always {} due to calling ServerReset (via _Reset()) immediately
prior to starting the server.

NOTE: LSP supports a didChangeConfiguration notification which we could
send should the settings change, but we don't currently so there is no
point repeatedly calling the user's Settings method.