Improve the project directory detection
In LSP's view of the world, everything has a project root directory.
Currently we dumbly use the path of the first file opened as the root
unless the completer implementation does something smarter. For example,
the java completer tries to find a project config file.

We now try and guess a project directory using the following hierarchy:
 - extra conf dir if it is not the global extra conf
 - client working dir
 - dir of first opened file

Other LSP servers legitimately expect a reasonable project directory to
be supplied on initialisation.