Improve cache in LSP completer
LSP allows servers to return an incomplete list of completions. The
cache cannot be used in that case and the current column must be sent to
the language server for the subsequent completion requests; otherwise,
the server will return the same incomplete list. When that list is
complete, two cases are considered:
 - the starting column was sent to the server: cache is valid for the
   whole completion;
 - the current column was sent to the server: cache stays valid while
   the cached query is a prefix of the subsequent queries.