Support LSP declaration, typeDefinition, and implementation
Add the GoToType and GoToImplementation commands to the LSP completer.
They are respectively mapped to the textDocument/typeDefinition and
textDocument/implementation LSP endpoints.
Update the GoToDeclaration command to be mapped to the
textDocument/declaration endpoint if the LSP server supports it;
otherwise, the command is mapped to textDocument/definition as before.
Change the GoTo command to jump to the declaration if the cursor is
already on a definition and vice versa if the LSP server supports
declaration.