Don't consider clangd 10 out of date.
The regex used to parse the clangd version only allows major numbers as
a single digit. Consequently, this results in a parsing failure for the
following version string:

```
    clangd version 10.0.0 (https://github.com/llvm/llvm-project.git bd68a052f292b5df7c5717bd880d796ac7507fc0)
```

This diff changes the regex to allow for two digit major version
numbers, and changes the comparison from a string based comparison to a
tuple/int based comparison to ensure that 10.0.0 > 7.0.0.