Support user-specified project directory in LSP servers
Store all settings returned from the Settings function

Allow extra conf to supply project directory. This is useful at least
for java with complex workspaces containing multiple .project files

For the output in debug info, we _could_ change to show all settings,
but for now, let's keep it as it was before.