Clean up python's sys path
Removing python-future from the extra conf was missed in the python2
support drop pull request.

Numpydoc is now appended to the sys.path to allow python to find the
system-wide numpydoc instead. This is because we're unlikely to update
numpydoc as the v0.9.0+ depend on sphinx and that would be rather insane
to keep around as a submodule.