Resolve individual completions on-demand
Previously we would do completion/resolve for every matching completion
candidate up front. This was also embarrassingly serial. For 100
candidates, we'd do 100 synchronous resolve requests. This can lead to
timeouts.  Resolve a single candidate and return it, as this is all Vim
can really handle

This leads to unusable performance on the Godot server (because for some
reason there are a lot of candidates and they take a _long_ time to
resolve serially).

So we implement an endpoint to resolve a completion item on-demand, and
change the resolve behaviour so that we only ever resolve 10 candidates
up-front.

The approach is to add a `'resolve'` key to the `extra_data` dictionary
which is an ID telling the client that this item needs to be resolved to
fill in the extra data documentation. If this key is not present,
resolve is not required.

Couple of things to note:

1. we resolve up to 10 _matching_ candidates (i.e. we still do
   FilterAndSortCandidates _first_)
2. we therefore still have a mixture of resolved and un-resolved
   candidates in the cache
3. we only allow resolving a candidate if there is a cache HIT for the
   request data. this means we can use the index into the full set of
   items as the "key" for what to resolve.
4. we don't resolve things multiple times. for candidates that are
   already resolved, we don't include the 'resolve' key which tells the
   client that it might need resolving.