Add a TCP bridge for client-to-client communication
This is a bit of a hack, but it's required for testing. Both the server
(generic_server) and ycmd want to be the client of the TCP session (i.e.
connect to some listening port). In order to connect the two together we
use 2 instances of nc (netcat) which listen on differnt ports and pipe
the data through some fifos to each other.

This is sufficient to bridge in the tests. If we needed this for a real
server, we would implement the TCPSingleStreamServer and use that (it's
an almost trivial extension of TCPSingleStreamConnection).