Use raw CPython API in PythonSupport.cpp
Using the raw C API allows access to objects internals without checking
for errors. This is fine, because ycmd already checks everything
necessary. Plus, pybind11 can easily fall back to generic
implementations - for example, `len()` is `PyObject_Len()`, even if we
know that the object is `PyListObject*`. The C API, thus, allows us to
exploit the knowledge of types that pybind11 doesn't have.