Add very simple documentSymbol supoport
We re-use the GoToSymbol handling for the "legacy" list-of-symbols type
response from documentSymbol and just populate the location list in the
same way that GoToSymbol does.

This is not exactly a drawer/document outline from a classic IDE but it
servers a purpose of getting a list of all the thigns int the file if
you're looking for something specific or just browsing.