Add more symbol information to GoTo requests
This allows the cliet to display something a bit nicer in a symbol
search dialog. For example, they kind and name can be used to pick a
syntax group to highlight the name.

In addition, supply the description when returning a single result from
python GoTo requests. This allows for searches which return only a
single respnose to still be displayed in a symbol search dialog.

These changes are required for the YCM client symbol search.